requirements.txt
pytest == 7.4.2
requests == 2.31.0
behave == 1.2.6
selenium == 4.12.0
allure-behave == 2.13.2
allure-pytest == 2.13.2